+++ kado_modules/doc/main/index.js
@@ -55,14 +55,14 @@ exports.versionList = (req,res) => {
       if(!result || !result.length){
         throw new Error('This project has no versions')
       }
-      res.render(res.locals._view.get('doc/versionList'),{
+      res.render('doc/versionList',{
         project: result[0].DocProject,
         versionList: result,
         _pageTitle: K._l.doc.doc_project_version + ' ' + K._l.list
       })
     })
     .catch((err) => {
-      res.render(res.locals._view.get('error'),{error: err.message})
+      res.render('error',{error: err.message})
     })
 }
 
@@ -93,7 +93,7 @@ exports.list = (req,res) => {
       })
     })
     .catch((err) => {
-      res.render(res.locals._view.get('error'),{error: err.message})
+      res.render('error',{error: err.message})
     })
 }
 
@@ -132,7 +132,7 @@ exports.entry = (req,res) => {
       if(!result) throw new Error('Document not found')
       result.content = K.base64js.fromByteArray(
         new Buffer(result.content,'utf-8'))
-      res.render(res.locals._view.get('doc/entry'),{
+      res.render('doc/entry',{
         doc: result,
         docList: docList,
         _pageTitle: result.title
@@ -140,6 +140,6 @@ exports.entry = (req,res) => {
     })
     .catch((err) => {
       if('Document not found' === err.message) res.status(404)
-      res.render(res.locals._view.get('error'),{error: err})
+      res.render('error',{error: err})
     })
 }